feat: prompt sign-in on first settings open when signed out
The first time a signed-out user opens the BetterTTV settings menu, show
a one-time prompt nudging them to sign in. It appears only once ever,
whether they sign in or dismiss it.

Extract openSignInModal and openSubscriptionUpgradeModal out of
ProRequiredState into the modal util so the sign-in flow can be reused
outside the Pro-toggle hook, then use openSignInModal for the settings
prompt. The prompt gates on the persisted access token rather than the
async-populated user so it never fires for a signed-in user whose
profile hasn't finished syncing yet.
